refactor: Encapsulate block date logic within `DeltaLakeClient`
